🚚 Organize onboarding components
