--- modules/ppcp-settings/resources/js/Components/ReusableComponents/Icons.js
@@ -1,5 +1,0 @@
-export { default as openSignup } from './Icons/open-signup';
-export { default as logoPayPal } from './Icons/logo-paypal';
-
-export const NOTIFICATION_SUCCESS = '✔️';
-export const NOTIFICATION_ERROR = '❌';
