+++ modules/ppcp-settings/resources/js/Components/ReusableComponents/Icons/index.js
@@ -1,0 +1,5 @@
+export { default as openSignup } from './open-signup';
+export { default as logoPayPal } from './logo-paypal';
+
+export const NOTIFICATION_SUCCESS = '✔️';
+export const NOTIFICATION_ERROR = '❌';
