+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/Steps/StepWelcome.js
@@ -1,14 +1,14 @@
 import { __ } from '@wordpress/i18n';
 import { Button } from '@wordpress/components';
 
-import PaymentMethodIcons from '../../ReusableComponents/PaymentMethodIcons';
-import Separator from '../../ReusableComponents/Separator';
-import AccordionSection from '../../ReusableComponents/AccordionSection';
-import OnboardingHeader from './Components/OnboardingHeader';
-import WelcomeDocs from './Components/WelcomeDocs';
-import AdvancedOptionsForm from './Components/AdvancedOptionsForm';
-import { CommonHooks } from '../../../data';
-import BusyStateWrapper from '../../ReusableComponents/BusyStateWrapper';
+import PaymentMethodIcons from '../../../ReusableComponents/PaymentMethodIcons';
+import Separator from '../../../ReusableComponents/Separator';
+import AccordionSection from '../../../ReusableComponents/AccordionSection';
+import { CommonHooks } from '../../../../data';
+import BusyStateWrapper from '../../../ReusableComponents/BusyStateWrapper';
+import OnboardingHeader from '../Components/OnboardingHeader';
+import WelcomeDocs from '../Components/WelcomeDocs';
+import AdvancedOptionsForm from '../Components/AdvancedOptionsForm';
 
 const StepWelcome = ( { setStep, currentStep } ) => {
 	const { storeCountry } = CommonHooks.useWooSettings();
