+++ modules/ppcp-settings/resources/js/Components/Screens/Onboarding/Steps/index.js
